extract auth from appsync
