+++ packages/appsync/bin/typegen
