+++ packages/appsync/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aws-amplify-hooks/appsync",
-  "version": "0.0.5",
+  "version": "0.0.6",
   "license": "MIT",
   "scripts": {
     "build": "rm -rf dist/ && tsc",
@@ -13,9 +13,9 @@
   "types": "dist/index.d.ts",
   "dependencies": {
     "@apollo/client": "^3.6.9",
+    "aws-amplify": "^4.3.26",
     "aws-appsync-auth-link": "^3.0.7",
     "aws-appsync-subscription-link": "^3.1.0",
-    "aws-amplify": "^4.3.26",
     "uuid": "8.3.2"
   },
   "devDependencies": {
